+++ packages/react-native-messaging-core/android/src/main/java/com/nabla/sdk/reactnative/messaging/core/models/Conversation.kt
@@ -5,24 +5,24 @@ import com.facebook.react.bridge.ReadableArray
 import com.facebook.react.bridge.ReadableMap
 import com.nabla.sdk.messaging.core.domain.entity.Conversation
 
-internal fun Conversation.toMap(): ReadableMap {
-    return Arguments.createMap().apply {
-        putString("id", id.stableId.toString())
-        putString("title", title)
-        putString("inboxPreviewTitle", inboxPreviewTitle)
-        putString("lastMessagePreview", lastMessagePreview)
-        putString("lastModified", lastModified.toString())
-        putInt("patientUnreadMessageCount", patientUnreadMessageCount)
-        putArray("providers", providersInConversation.toMapArray())
-    }
-}
-
 @JvmName("toConversationMapArray")
 internal fun List<Conversation>.toMapArray(): ReadableArray {
     return Arguments.createArray().apply {
         forEach { conversation ->
             pushMap(conversation.toMap())
         }
+    }
+}
+
+internal fun Conversation.toMap(): ReadableMap {
+    return Arguments.createMap().apply {
+        putMap("id", id.toMap())
+        putString("title", title)
+        putString("inboxPreviewTitle", inboxPreviewTitle)
+        putString("lastMessagePreview", lastMessagePreview)
+        putString("lastModified", lastModified.toString())
+        putInt("patientUnreadMessageCount", patientUnreadMessageCount)
+        putArray("providers", providersInConversation.toMapArray())
     }
 }