+++ packages/react-native-messaging-core/src/types/ConversationId.ts
@@ -1,0 +1,3 @@
+import { LocalId, RemoteId } from './TransientId';
+
+export type ConversationId = RemoteId | LocalId;