+++ packages/react-native-messaging-core/src/types/MessageId.ts
@@ -1,0 +1,3 @@
+import { LocalId, RemoteId } from './TransientId';
+
+export type MessageId = RemoteId | LocalId;